lambda_function/requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.1.1
Deprecated==1.2.14
frozenlist==1.4.1
gitdb==4.0.11
github==1.2.7
GitPython==3.1.42
idna==3.6
multidict==6.0.5
pycparser==2.21
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
smmap==5.0.1
typing_extensions==4.9.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
